reddit_server/requirements.txt
# MCP Server dependencies
mcp>=0.9.0

# Reddit API
praw>=7.7.1
asyncpraw>=7.7.1

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
aiohttp>=3.9.0
aiofiles>=23.0.0

# Data handling
python-dateutil>=2.8.2

# Logging
colorlog>=6.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0